removed extra type conversion

+++ ignite/metrics/gan/fid.py
@@ -183,8 +183,8 @@ class FID(_BaseInceptionMetric):
 
         train, test = output
 
-        train_features = self._extract_features(train).to(torch.float64)
-        test_features = self._extract_features(test).to(torch.float64)
+        train_features = self._extract_features(train)
+        test_features = self._extract_features(test)
 
         if train_features.shape[0] != test_features.shape[0] or train_features.shape[1] != test_features.shape[1]:
             raise ValueError(
